requirements.txt
# 多平台视频搜索API依赖

# ============ Web框架 ============
fastapi==0.109.0
uvicorn[standard]==0.27.0

# ============ 数据验证 ============
pydantic==2.5.3
pydantic-settings==2.1.0

# ============ 环境变量 ============
python-dotenv==1.0.0

# ============ HTTP请求 ============
requests==2.31.0
httpx==0.26.0

# ============ 浏览器自动化（抖音/小红书搜索需要） ============
# MediaCrawler 桥接脚本依赖
playwright==1.45.0
Pillow==10.2.0
tenacity==8.2.2
aiofiles~=23.2.1
parsel==1.9.1

# ============ 工具库 ============
